Fix inline function 'void TglImpl::RendererImpl::Destroy()' used but never defined
